[Integration Tests] Improve and fix integration tests
 ### Motivation

Integration test are basically just not working in apache CI. there are multiple reasons:

1. PULSAR_MEM is set to very high (~2G). If we start more than 6 containers, it will quickly go up to ~12GB.
2. We start containers for each tests, which is very inefficient.
3. There is a regression from #2214. #2214 tried to listen on a hostname that is configured in worker config,
  which the hostname is internally to docker environment.

 ### Changes

1. Set PULSAR_MEM to use no more than 128M.
2. Switch to use test suite, trying to start containers only once as possible and use them across test suites.
   so we only start the cluster for a set of tests. hence reorganize those tests and get rid
   of using dataProvider, which doesn't work well with current approach.
3. Remove binding to hostname in WorkerServer, so it binds to all interfaces.
