+++ tests/integration/pom.xml
@@ -102,6 +102,9 @@
             <currentVersion>${project.version}</currentVersion>
             <maven.buildDirectory>${project.build.directory}</maven.buildDirectory>
           </systemPropertyVariables>
+          <suiteXmlFiles>
+            <file>src/test/resources/suite.xml</file>
+          </suiteXmlFiles>
         </configuration>
       </plugin>
     </plugins>
@@ -131,6 +134,12 @@
               -Dio.netty.leakDetectionLevel=advanced
               </argLine>
               <skipTests>false</skipTests>
+              <suiteXmlFiles>
+                <file>src/test/resources/pulsar-process.xml</file>
+                <file>src/test/resources/pulsar-thread.xml</file>
+                <file>src/test/resources/tiered-storage.xml</file>
+              </suiteXmlFiles>
+              <forkCount>1</forkCount>
             </configuration>
           </plugin>
         </plugins>
